pyproject.toml
[project]
name = "scrap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "beautifulsoup4>=4.13.3",
    "fastapi[standart]>=0.115.12",
    "notebook>=7.3.3",
    "openai>=1.72.0",
    "pandas>=2.2.3",
    "playwright>=1.51.0",
    "playwright-stealth>=1.0.6",
    "playwrightcapture[recaptcha]>=1.28.5",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
